# apis/app-api/store/openapi.yaml
openapi: 3.1.2
info:
  title: SDKWork App Store App API
  version: 1.0.0
  description: 'App/client contract for SDKWork App Store catalog, listings, publishers,
    releases, compliance, library, and wishlist capabilities.

    '
  x-sdkwork-owner: sdkwork-appstore
  x-sdkwork-api-authority: sdkwork-appstore-app-api
  x-sdkwork-sdk-family: sdkwork-appstore-app-sdk
  x-sdkwork-domain: appstore
  x-sdkwork-standard-profile: sdkwork-appstore-app-v3
  x-sdkwork-audience: Store clients, publisher console, mobile/H5/PC apps
servers:
- url: http://127.0.0.1:18090
  description: Local sdkwork-appstore runtime
...
security:
- AuthToken: []
  AccessToken: []
paths:
...
  /app/v3/api/listings:
    post:
      tags:
      - listings
      summary: Create listing for registered app
      operationId: appstore.listings.create
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SdkWorkListResponse'
  /app/v3/api/listings/{listingId}:
    get:
      tags:
      - listings
      summary: Retrieve listing detail
      operationId: appstore.listings.retrieve
      parameters:
      - $ref: '#/components/parameters/ListingId'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SdkWorkListResponse'
    patch:
      tags:
      - listings
      summary: Update listing metadata
      operationId: appstore.listings.update
      parameters:
      - $ref: '#/components/parameters/ListingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingUpdateRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SdkWorkListResponse'
  /app/v3/api/listings/{listingId}/localizations/{locale}:
    put:
      tags:
      - listings
      summary: Upsert listing localization
      operationId: appstore.listings.localization.upsert
      parameters:
      - $ref: '#/components/parameters/ListingId'
      - $ref: '#/components/parameters/Locale'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingLocalizationUpsertRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SdkWorkListResponse'
  /app/v3/api/listings/{listingId}/media:
    get:
      tags:
      - listings
      summary: List listing media
      operationId: appstore.listings.media.list
      parameters:
      - $ref: '#/components/parameters/ListingId'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SdkWorkListResponse'
    post:
      tags:
      - listings
      summary: Attach listing media
      operationId: appstore.listings.media.attach
      parameters:
      - $ref: '#/components/parameters/ListingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingMediaAttachRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SdkWorkListResponse'
  /app/v3/api/listings/{listingId}/media/{mediaId}:
    delete:
      tags:
      - listings
      summary: Remove listing media
      operationId: appstore.listings.media.remove
      parameters:
      - $ref: '#/components/parameters/ListingId'
      - name: mediaId
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Removed
  /app/v3/api/listings/{listingId}/categories:
    put:
      tags:
      - listings
      summary: Bind listing categories
      operationId: appstore.listings.categories.bind
      parameters:
      - $ref: '#/components/parameters/ListingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingCategoryBindRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SdkWorkListResponse'
  /app/v3/api/listings/{listingId}/regions:
    put:
      tags:
      - listings
      summary: Update regional availability
      operationId: appstore.listings.regions.update
      parameters:
      - $ref: '#/components/parameters/ListingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegionalAvailabilityUpdateRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SdkWorkListResponse'
  /app/v3/api/listings/{listingId}/releases:
    get:
      tags:
      - listings
      summary: List releases for listing
      operationId: appstore.listings.releases.list
      parameters:
      - $ref: '#/components/parameters/ListingId'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SdkWorkListResponse'
    post:
      tags:
      - releases
      summary: Create release for listing
      operationId: appstore.releases.create
      parameters:
      - $ref: '#/components/parameters/ListingId'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReleaseCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/SdkWorkApiResponse'
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      type: object
                      required:
                      - item
                      properties:
                        item:
                          $ref: '#/components/schemas/ReleaseResponse'
  /app/v3/api/listings/{listingId}/releases/history:
    get:
      tags:
      - listings
      summary: List release history for listing
      operationId: appstore.listings.releases.history.list
      parameters:
      - $ref: '#/components/parameters/ListingId'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SdkWorkListResponse'
  /app/v3/api/listings/{listingId}/similar:
    get:
      tags:
      - listings
      summary: List similar listings
      operationId: appstore.listings.similar.list
      parameters:
      - $ref: '#/components/parameters/ListingId'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SdkWorkListResponse'
  /app/v3/api/listings/{listingId}/developer_other:
    get:
      tags:
      - listings
      summary: List other listings from the same developer
      operationId: appstore.listings.developerOther.list
      parameters:
      - $ref: '#/components/parameters/ListingId'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SdkWorkListResponse'
  /app/v3/api/listings/{listingId}/editorial:
    get:
      tags:
      - listings
      summary: Retrieve listing editorial content
      operationId: appstore.listings.editorial.retrieve
      parameters:
      - $ref: '#/components/parameters/ListingId'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SdkWorkResourceResponse'
  /app/v3/api/listings/{listingId}/submissions:
    post:
      tags:
      - listings
      summary: Submit listing for review
      operationId: appstore.listings.submissions.create
      parameters:
      - $ref: '#/components/parameters/ListingId'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingSubmissionCreateRequest'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SdkWorkListResponse'
...
components:
  securitySchemes:
    AuthToken:
      type: apiKey
      in: header
      name: Auth-Token
    AccessToken:
      type: apiKey
      in: header
      name: Access-Token
  parameters:
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
    ListingId:
      name: listingId
      in: path
      required: true
      schema:
        type: string
...
    Locale:
      name: locale
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    ListingCreateRequest:
      type: object
      required:
      - appId
      - appKey
      - publisherId
      - defaultLocale
      properties:
        appId:
          type: string
        appKey:
          type: string
        publisherId:
          type: string
        listingSlug:
          type: string
        pricingModel:
          type: string
        defaultLocale:
          type: string
...
    ListingUpdateRequest:
      type: object
      properties:
        pricingModel:
          type: string
        officialWebsiteUrl:
          type: string
        supportUrl:
          type: string
        privacyPolicyUrl:
          type: string
...
    ListingLocalizationUpsertRequest:
      type: object
      required:
      - displayName
      - shortDescription
      - fullDescription
      properties:
        displayName:
          type: string
        subtitle:
          type: string
        shortDescription:
          type: string
        fullDescription:
          type: string
        keywords:
          type: array
          items:
            type: string
...
    ListingMediaAttachRequest:
      type: object
      required:
      - mediaRole
      - mediaResourceId
      properties:
        mediaRole:
          type: string
          enum:
          - ICON
          - SCREENSHOT
          - PREVIEW_VIDEO
          - FEATURE_GRAPHIC
        mediaResourceId:
          type: string
        platformScope:
          type: string
        locale:
          type: string
...
    ListingCategoryBindRequest:
      type: object
      required:
      - categoryIds
      properties:
        primaryCategoryId:
          type: string
        categoryIds:
          type: array
          items:
            type: string
    RegionalAvailabilityUpdateRequest:
      type: object
      required:
      - regions
      properties:
        regions:
          type: array
          items:
            type: object
            required:
            - regionCode
            - availabilityStatus
            properties:
              regionCode:
                type: string
              availabilityStatus:
                type: string
...
    ListingSubmissionCreateRequest:
      type: object
      required:
      - submissionType
      properties:
        submissionType:
          type: string
          enum:
          - INITIAL
          - METADATA
          - RELEASE
        releaseId:
          type: string
...
    ReleaseCreateRequest:
      type: object
      required:
      - channelCode
      - versionName
      - versionCode
      properties:
        channelCode:
          type: string
        versionName:
          type: string
        versionCode:
          type: string
        buildNumber:
          type: string
...
    ReleaseResponse:
      $ref: '#/components/schemas/SdkWorkResourceResponse'
...
    SdkWorkApiResponse:
      type: object
      additionalProperties: false
      required:
      - code
      - data
      - traceId
      properties:
        code:
          type: integer
          format: int32
          enum:
          - 0
          default: 0
          minimum: 0
          maximum: 0
        data:
          description: Operation-specific payload typed per response schema.
        traceId:
          type: string
          format: uuid
          description: Server-owned request correlation id.
    SdkWorkResourceData:
      type: object
      additionalProperties: false
      required:
      - item
      properties:
        item:
          type: object
          additionalProperties: true
          description: Typed domain resource for the operation.
    SdkWorkPageData:
      type: object
      additionalProperties: false
      required:
      - items
      - pageInfo
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: true
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
...
    PageInfo:
      type: object
      additionalProperties: false
      required:
      - mode
      properties:
        mode:
          type: string
          enum:
          - offset
          - cursor
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 200
        totalItems:
          type: string
          pattern: ^[0-9]+$
        totalPages:
          type: integer
          minimum: 0
        nextCursor:
          type:
          - string
          - 'null'
        hasMore:
          type: boolean
...
    SdkWorkResourceResponse:
      allOf:
      - $ref: '#/components/schemas/SdkWorkApiResponse'
      - type: object
        required:
        - data
        properties:
          data:
            $ref: '#/components/schemas/SdkWorkResourceData'
    SdkWorkListResponse:
      allOf:
      - $ref: '#/components/schemas/SdkWorkApiResponse'
      - type: object
        required:
        - data
        properties:
          data:
            $ref: '#/components/schemas/SdkWorkPageData'
...